backend/requirements.txt
fastapi==0.115.10
uvicorn[standard]==0.34.0
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1
python-multipart==0.0.20
requests==2.32.3
python-dotenv==1.0.1
google-genai==0.7.0
pypdf==5.3.1
python-docx==1.1.2
python-pptx==1.0.2
